requirements.txt
torch>=1.10.0
transformers>=4.25.0
pandas>=1.5.0
tqdm>=4.64.0
Pillow>=9.3.0
clip@git+https://github.com/openai/CLIP.git